% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARCH.Test.R
\name{ARCH.Test}
\alias{ARCH.Test}
\title{ARCH Test for time series}
\usage{
ARCH.Test(x, lags)
}
\arguments{
\item{x}{residual series name,}

\item{lags}{lags}
}
\description{
This function allows you to make ARCH Test for residuals
}
\examples{
set.seed(12345)
x <- rnorm(1000)
ARCH.Test(x,3)

}
\keyword{ARCH}
\keyword{test}
