% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cook_Vougas_2009_unit_root.R
\name{Cook_Vougas_2009_unit_root}
\alias{Cook_Vougas_2009_unit_root}
\title{Cook and Vougas(2009) nonlinear unit root test function}
\usage{
Cook_Vougas_2009_unit_root(x, model, max_lags)
}
\arguments{
\item{x}{series name,}

\item{model}{if model A 1, if model B 2, if model C 3, model D 4}

\item{max_lags}{maximum lag(optimal lag selected by AIC)}
}
\description{
This function allows you to make Cook and Vougas(2009) nonlinear unit root test
}
\examples{
\donttest{
set.seed(12345)
x <- rnorm(1000)
Cook_Vougas_2009_unit_root(x,model=1,max_lags=3)

data(IBM)
Cook_Vougas_2009_unit_root(x=IBM,model=3,max_lags=3)

}

}
\keyword{nonlinear}
\keyword{root}
\keyword{test}
\keyword{unit}
