% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM.R
\docType{data}
\name{IBM}
\alias{IBM}
\title{IBM}
\format{
A data frame containing :

Price IBM Close Price
}
\source{
Yahoo Finance
}
\usage{
IBM
}
\description{
Daily time series data between 01.01.2010 - 01.01.2018
}
\examples{
summary(IBM)
}
\keyword{datasets}
