% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketPrices.R
\docType{data}
\name{MarketPrices}
\alias{MarketPrices}
\title{MarketPrices}
\format{
A data frame containing :

FCHI CAC 40 Paris Stock Exchange Prices

IBEX Madrid Stock Exchange Prices
}
\source{
Yahoo Finance
}
\usage{
MarketPrices
}
\description{
Daily time series data between 01.01.2014-01.01.2019
}
\examples{
summary(MarketPrices)
}
\keyword{datasets}
