% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McLeodLi.R
\name{Mc.Leod.Li}
\alias{Mc.Leod.Li}
\title{Mc.Leod.Li nonlinearity test}
\usage{
Mc.Leod.Li(y, lag)
}
\arguments{
\item{y}{series name,}

\item{lag}{lag parameter,}
}
\value{
"lag stat pvalue" the lag order, the value of the test statistic and the probability of test statistic, respectively.
}
\description{
This function allows you to make Mc.Leod.Li nonlinearity test
}
\examples{
x <- rnorm(1000)
Mc.Leod.Li(x, 10)


data(IBM)
Mc.Leod.Li(IBM,4)


}
\references{
Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{nonlinearity}
\keyword{test}
