% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETAR_model.R
\name{SETAR_model}
\alias{SETAR_model}
\title{SETAR model estimation}
\usage{
SETAR_model(y, delay_order, lag_length, trim_value)
}
\arguments{
\item{y}{series name,}

\item{delay_order}{Delay order,}

\item{lag_length}{lag length}

\item{trim_value}{trimmed value, .15, .10, .5}
}
\value{
"Model" Estimated model

"threshold" the value of threshold
}
\description{
This function allows you to estimate SETAR model
}
\examples{

\donttest{
x <- rnorm(100)
SETAR_model(x, 1, 12, .15)


data(IBM)
SETAR_model(IBM, 1, 12, .05)
}


}
\references{
Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{estimation}
\keyword{model}
\keyword{nonlinear}
